feature(pagination): move loader component to utils
