+++ client/components/auth/ForgotPasswordForm.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 import Header from '../Header';
-import Loading from '../Loading';
+import Loading from '../common/Loading';
 
 const ForgotPasswordForm = props => (
   <div>
