+++ client/components/auth/ResetPasswordForm.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { Row, Col } from 'react-materialize';
 
 import Header from '../Header';
-import Loading from '../Loading';
+import Loading from '../common/Loading';
 
 
 const ResetPasswordForm = props => (
