+++ client/utils/saveLocally.js
@@ -9,6 +9,7 @@ export const saveState = (state) => {
     localStorage.setItem('state', serializedState);
   } catch (e) {
     console.log(e);
+    return false;
   }
 };
 
@@ -20,10 +21,11 @@ export const loadState = () => {
   try {
     const serializedState = localStorage.getItem('state');
     if (serializedState === null) {
-      return undefined;
+      return false;
     }
     return JSON.parse(serializedState);
   } catch (e) {
     console.log(e);
+    return false;
   }
 };
